% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_relationships}
\alias{as_relationships}
\title{iris %>%
  filter(Species == "setosa") %>%
  as_nodes(label = Species) %>%
  call_api(con, include_stats = TRUE)}
\usage{
as_relationships(tbl, start_node_label, start_node_properties, relationships,
  end_node_label, end_node_properties)
}
\description{
iris %>%
  filter(Species == "setosa") %>%
  as_nodes(label = Species) %>%
  call_api(con, include_stats = TRUE)
}
