% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to.R
\name{convert_to}
\alias{convert_to}
\title{Convert to another graph format}
\usage{
convert_to(res, format = c("visNetwork", "igraph"), label = name)
}
\arguments{
\item{res}{the res object from the API}

\item{format}{the format (either igraph or visNetwork)}

\item{label}{the column to be considered as the label columns}
}
\value{
a transformed list
}
\description{
Convert to another graph format
}
