% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract_nodes}
\alias{extract_nodes}
\alias{extract_relationships}
\title{Extract nodes or relationships}
\usage{
extract_nodes(x)

extract_relationships(x)
}
\arguments{
\item{x}{a result from Neo4J}
}
\value{
a tibble
}
\description{
Extract nodes or relationships
}
