% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_pane_rstudio.R
\name{launch_con_pane}
\alias{launch_con_pane}
\title{Launch Neo4J Connection Pane}
\usage{
launch_con_pane(con)
}
\arguments{
\item{con}{a connection object}
}
\value{
an opened Connection Pane
}
\description{
Launch Neo4J Connection Pane
}
