% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\docType{data}
\name{neo4j_api}
\alias{neo4j_api}
\title{A Neo4J Connexion}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
neo4j_api
}
\value{
A Neo4J Connexion
}
\description{
A Neo4J Connexion
}
\section{Methods}{

\describe{
\item{\code{access}}{list url, user and password}
\item{\code{ping}}{test your connexion}
\item{\code{version}}{Neo4J version}
\item{\code{get}}{Get a list of either relationship, labels, }
\item{\code{get}}{Get a list of either relationship, labels, }
\item{\code{get}}{Get a list of either relationship, labels, }
\item{\code{get}}{Get a list of either relationship, labels, }
\item{\code{get}}{Get a list of either relationship, labels, }
}
}

\section{Data}{

\describe{
\item{\code{url}}{list url, user and password}
\item{\code{user}}{test your connexion}
}
}

\examples{
\dontrun{
con <- neo4j_api$new(url = "http://localhost:7474", user = "neo4j", password = "password")
}


}
\keyword{datasets}
