% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_api.R
\name{neo4r_cypher}
\alias{neo4r_cypher}
\title{Run A Cypher Query (but start with the connection object)}
\usage{
neo4r_cypher(con, query, params, ...)
}
\arguments{
\item{con}{A NEO4JAPI connection object}

\item{query}{The cypher query}

\item{params}{A list of parameters to pass along with the cypher query}

\item{...}{arguments to pass along to the call_neo4j function}
}
\value{
the result from the Neo4j Call
}
\description{
Run A Cypher Query (but start with the connection object)
}
