% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_pane_rstudio.R
\name{on_connection_opened}
\alias{on_connection_opened}
\title{Opening the Neo4J connection Pane}
\usage{
on_connection_opened(con, url, auth)
}
\arguments{
\item{con}{A connection object}

\item{url}{The url of the database}

\item{auth}{The auth token}
}
\description{
For intenral use
}
\keyword{internal}
