% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cypher.R
\name{read_cypher}
\alias{read_cypher}
\title{Read a cypher file}
\usage{
read_cypher(file)
}
\arguments{
\item{file}{the path to the cypher file}
}
\value{
a tibble with the queries
}
\description{
Read a cypher file
}
\examples{
\dontrun{
read_cypher("random/create.cypher")
}
}
