% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cypher.R
\name{send_cypher}
\alias{send_cypher}
\title{Send a cypher file to be executed}
\usage{
send_cypher(
  path,
  con,
  type = c("row", "graph"),
  output = c("r", "json"),
  include_stats = TRUE,
  meta = FALSE
)
}
\arguments{
\item{path}{the path to the cypher file}

\item{con}{a connexion object created with neo4j_api$new()}

\item{type}{the type of the format to query for (row or graph)}

\item{output}{the printing method (r or json)}

\item{include_stats}{whether of not to include stats}

\item{meta}{whether of not to include meta info}
}
\value{
a cypher call
}
\description{
Send a cypher file to be executed
}
\examples{
\dontrun{
send_cypher("random/create.cypher")
path <- "data-raw/constraints.cypher"
}
}
