% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_them.R
\name{unnest_graph}
\alias{unnest_graph}
\title{Unnest both relationships and nodes}
\usage{
unnest_graph(res)
}
\arguments{
\item{res}{an api graph result}
}
\value{
a list of two unnested data.frames
}
\description{
Unnest both relationships and nodes
}
