% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_them.R
\name{unnest_nodes}
\alias{unnest_nodes}
\title{Unnest a node data.frame}
\usage{
unnest_nodes(nodes_tbl, what = c("all", "label", "properties"))
}
\arguments{
\item{nodes_tbl}{the node table}

\item{what}{what to unnest}
}
\value{
a new dataframe
}
\description{
Unnest a node data.frame
}
