% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_them.R
\name{unnest_relationships}
\alias{unnest_relationships}
\title{Unnest a Relationships table}
\usage{
unnest_relationships(relationships_tbl)
}
\arguments{
\item{relationships_tbl}{a relationship table}
}
\value{
an unnested table
}
\description{
Unnest a Relationships table
}
