% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{neo4j_api}
\alias{neo4j_api}
\title{A Neo4J Connexion}
\value{
A Neo4J Connexion
}
\description{
A Neo4J Connexion

A Neo4J Connexion
}
\section{Methods}{

\describe{
\item{\code{access}}{list url,db, user and password}
\item{\code{ping}}{test your connexion}
\item{\code{version}}{Neo4J version}
\item{\code{get}}{Get a list of either relationship, labels, }
\item{\code{get}}{Get a list of either relationship, labels, }
\item{\code{get}}{Get a list of either relationship, labels, }
\item{\code{get}}{Get a list of either relationship, labels, }
\item{\code{get}}{Get a list of either relationship, labels, }
}
}

\section{Data}{

\describe{
\item{\code{url}}{list url, user and password}
\item{\code{user}}{test your connexion}
}
}

\examples{
\dontrun{
con <- neo4j_api$new(url = "http://localhost:7474", db = "neo4j", user = "neo4j", password = "password")
}


}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{neo4j_api$print()}}
\item \href{#method-new}{\code{neo4j_api$new()}}
\item \href{#method-reset_url}{\code{neo4j_api$reset_url()}}
\item \href{#method-reset_user}{\code{neo4j_api$reset_user()}}
\item \href{#method-reset_password}{\code{neo4j_api$reset_password()}}
\item \href{#method-access}{\code{neo4j_api$access()}}
\item \href{#method-ping}{\code{neo4j_api$ping()}}
\item \href{#method-get_version}{\code{neo4j_api$get_version()}}
\item \href{#method-get_relationships}{\code{neo4j_api$get_relationships()}}
\item \href{#method-get_labels}{\code{neo4j_api$get_labels()}}
\item \href{#method-get_property_keys}{\code{neo4j_api$get_property_keys()}}
\item \href{#method-get_index}{\code{neo4j_api$get_index()}}
\item \href{#method-get_constraints}{\code{neo4j_api$get_constraints()}}
\item \href{#method-clone}{\code{neo4j_api$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$new(url, db, user, password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_url"></a>}}
\if{latex}{\out{\hypertarget{method-reset_url}{}}}
\subsection{Method \code{reset_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$reset_url(url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_user"></a>}}
\if{latex}{\out{\hypertarget{method-reset_user}{}}}
\subsection{Method \code{reset_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$reset_user(user)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_password"></a>}}
\if{latex}{\out{\hypertarget{method-reset_password}{}}}
\subsection{Method \code{reset_password()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$reset_password(password)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-access"></a>}}
\if{latex}{\out{\hypertarget{method-access}{}}}
\subsection{Method \code{access()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$access()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ping"></a>}}
\if{latex}{\out{\hypertarget{method-ping}{}}}
\subsection{Method \code{ping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$ping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_version"></a>}}
\if{latex}{\out{\hypertarget{method-get_version}{}}}
\subsection{Method \code{get_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$get_version()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_relationships"></a>}}
\if{latex}{\out{\hypertarget{method-get_relationships}{}}}
\subsection{Method \code{get_relationships()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$get_relationships()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_labels"></a>}}
\if{latex}{\out{\hypertarget{method-get_labels}{}}}
\subsection{Method \code{get_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$get_labels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_property_keys"></a>}}
\if{latex}{\out{\hypertarget{method-get_property_keys}{}}}
\subsection{Method \code{get_property_keys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$get_property_keys()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_index"></a>}}
\if{latex}{\out{\hypertarget{method-get_index}{}}}
\subsection{Method \code{get_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$get_index()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_constraints"></a>}}
\if{latex}{\out{\hypertarget{method-get_constraints}{}}}
\subsection{Method \code{get_constraints()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$get_constraints()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{neo4j_api$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
